[Update] 判斷若找不到user id / post id 則回傳錯誤訊息

+++ routes/index.js
@@ -49,22 +49,27 @@ router.get("/posts", async (req, res) => {
 router.post("/posts", async (req, res) => {
   try {
     const data = req.body;
-
-    if (data.title !== undefined) {
-      const post = {
-        userData: data.userId,
-        title: data.title,
-        content: data.content,
-        imgUrl: data.imgUrl,
-        tag: data.tag,
-      };
-      await Post.create(post).then(async () => {
-        const allData = await Post.find();
-        successHandler(res, allData);
-      });
-    } else {
-      errorHandler(res, "POST title未填寫");
+    const user = await User.findById(data.userId);
+    if (user === null) {
+      errorHandler(res, "此User不存在");
+      return;
     }
+    if (data.title === undefined) {
+      errorHandler(res, "POST title未填寫");
+      return;
+    }
+
+    const post = {
+      userData: data.userId,
+      title: data.title,
+      content: data.content,
+      imgUrl: data.imgUrl,
+      tag: data.tag,
+    };
+    await Post.create(post).then(async () => {
+      const allData = await Post.find();
+      successHandler(res, allData);
+    });
   } catch (error) {
     console.log("req error", error);
     errorHandler(res, `POST 格式錯誤，${error}`);
@@ -84,6 +89,11 @@ router.delete("/posts", async (req, res) => {
 
 router.delete("/posts/:id", async (req, res) => {
   const id = req.params.id;
+  const findPostId = await Post.findById(id);
+  if (findPostId === null) {
+    errorHandler(res, "文章不存在");
+    return;
+  }
   await Post.findByIdAndDelete(id)
     .then(async () => {
       const allData = await Post.find();
@@ -99,25 +109,31 @@ router.patch("/posts/:id", async (req, res) => {
     const data = req.body;
     const id = req.params.id;
 
-    if (data.title !== undefined) {
-      const post = {
-        tag: data.tag,
-        userName: data.userName,
-        title: data.title,
-        content: data.content,
-        imgUrl: data.imgUrl,
-      };
-      await Post.findByIdAndUpdate(id, post)
-        .then(async () => {
-          const allData = await Post.find();
-          successHandler(res, allData);
-        })
-        .catch(() => {
-          errorHandler(res, error);
-        });
-    } else {
-      errorHandler(res, "格式錯誤或無該筆資料");
+    const findUserId = await Post.findById(id);
+    if (findUserId === null) {
+      errorHandler(res, "Post不存在");
+      return;
     }
+    if (data.title === undefined) {
+      errorHandler(res, "格式錯誤或無該筆資料");
+      return;
+    }
+
+    const post = {
+      tag: data.tag,
+      userName: data.userName,
+      title: data.title,
+      content: data.content,
+      imgUrl: data.imgUrl,
+    };
+    await Post.findByIdAndUpdate(id, post)
+      .then(async () => {
+        const allData = await Post.find();
+        successHandler(res, allData);
+      })
+      .catch(() => {
+        errorHandler(res, error);
+      });
   } catch (error) {
     errorHandler(res, "格式錯誤");
   }
